Create README.md
Abstract


This document is a work-in-progress report about replicating results of Black Widow optimization. Our goal is to implement the pseudo-code by Hayyolalam, Kazem [2020] in Python and test its results on the sphere function. We will compare the results of our version vis-à-vis other versions of the same algorithm. First, we explain what metaheuristics are and the Black Widow metaphor. After explaining evolutive metaheuristics in general and Black Widow optimization in particular, we explain our progress. So far, we have programmed the sphere function and some components of the optimization model such as the iterator of generations, mating, the male/female distinction, and father cannibalism.